+++ test/com/walmartlabs/lacinia/parser/query_test.clj
@@ -1,5 +1,6 @@
 (ns com.walmartlabs.lacinia.parser.query-test
   (:require [clojure.test :refer [deftest is testing]]
+            [com.walmartlabs.test-utils :refer [compile-schema execute]]
             [com.walmartlabs.test-schema :refer [test-schema]]
             [com.walmartlabs.lacinia.schema :as schema]
             [com.walmartlabs.lacinia.parser :as parser]))
@@ -51,3 +52,13 @@
     (is (= {:id "1138"
             :newHomePlanet "❄ＨＯＴＨ❄"}
            (args "mutation { changeHeroHomePlanet(id: \"1138\", newHomePlanet: \"\\u2744\\uff28\\uff2f\\uff34\\uff28\\u2744\") {name}}")))))
+
+(deftest query-reserved-word
+  ;; Use 'query', 'mutation', and 'subscription' in various unusual places.
+  (let [schema (compile-schema "query-reserved.edn"
+                               {:resolve-query (fn [_ args _]
+                                                 args)})
+
+        result (execute schema "{ query(mutation: true) { mutation }}")]
+    (is (= {:data {:query {:mutation true}}}
+           result))))